requirements.txt
Pillow>=8.2.0
torch>=1.7.0
matplotlib>=3.1.0
torchvision>=0.8.0
